requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\Users\1998a\Documents\GitHub\first_st\first_st.py: 3
Pillow == 7.2.0

# C:\Users\1998a\Documents\GitHub\first_st\first_st.py: 2
pandas == 1.0.5

# C:\Users\1998a\Documents\GitHub\first_st\first_st.py: 1
streamlit == 0.74.1
